Rename all `votingResult.{yes,no}` attrs to `{in_favor,against}`
These are client renames to match the api changes.

I updated the `fetch_schema.yml` cassette by hand, rather than try to get a server up on the new api
and capture it that way. Since that is very hard to edit, we should regenerate (by deleting and running
the specs) the cassette once we've deployed the new api.

Part of liquidvotingio/api#46